% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_icon_color.R
\name{change_icon_color}
\alias{change_icon_color}
\title{Change Icon Color}
\usage{
change_icon_color(icon, color)
}
\arguments{
\item{icon}{icon, which is a 4D array}

\item{color}{color to change, in text form, passed to \code{\link{col2rgb}}}
}
\value{
A 4D array of the icon
}
\description{
Change Icon Color
}
\examples{
icon = scifigure::icons[[2]]
color = "blue"
original_color = "red"
icon2 = change_icon_color(icon, color)
}
