% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icons_diff}
\alias{icons_diff}
\title{repfigure icons_diff}
\format{
A list of length 44, with each item a 75x75x4 bitmap
}
\usage{
icons_diff
}
\description{
A dataset containing icon images showing difference rather than entity
used to render all difference figures in the scifigure package.
}
\keyword{datasets}
