% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_experiments.R
\name{init_experiments}
\alias{init_experiments}
\title{Initialize a skeleton data frame to create a figure with \code{sci_figure}}
\usage{
init_experiments(
  nexp = 3,
  exp_names = paste0("Exp", 1:nexp),
  stage_names = c("population", "question", "hypothesis", "experimental_design",
    "experimenter", "data", "analysis_plan", "analyst", "code", "estimate", "claim")
)
}
\arguments{
\item{nexp}{The number of scientific experiments to be represented in the data frame, i.e. number of columns.}

\item{exp_names}{The names of each experiment, i.e. column names. Default: "Exp1, Exp2, ..."}

\item{stage_names}{The names of each step in the process, i.e. row names. Defaults match Patil et. al.}
}
\description{
\code{init_experiments} generates a dataframe with the proper row and
column headers for user manipulation before calling \code{sci_figure}
}
\examples{
# Generate the default data frame of three experiments
init_experiments()

init_experiments(nexp = 5,
exp_names = c("Run_16_01", "Run_16_04", "Run_16_07",
	"Run_16_09", "Run_16_12"))
testthat::expect_error({
init_experiments(nexp = 2, exp_names = names)
})
}
\seealso{
\code{\link{sci_figure}}
}
