% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfigure.R
\name{repfigure}
\alias{repfigure}
\title{Create a figure depicting reproducibility/replicability}
\usage{
repfigure(type = "repl")
}
\arguments{
\item{type}{Either "repl" for replication or "repr" for reproduction}
}
\description{
\code{repfigure} creates a graphical representation of changes in a
second study or reproduction attempt as compared to an original study.
}
\note{
Work in progress...going to add user input
}
\examples{
repfigure("repl")
repfigure("repr")

}

