% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_figure.R
\name{replicate_figure}
\alias{replicate_figure}
\title{Create a figure depicting replicability}
\usage{
replicate_figure()
}
\description{
\code{replicate_figure} is a wrapper around the \code{sci_figure} function
to illustrate replicability in a two-experiment setting.
}

