% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_figure.R
\name{reproduce_figure}
\alias{reproduce_figure}
\title{Create a figure depicting reproducibility}
\usage{
reproduce_figure()
}
\description{
\code{reproduce_figure} is a wrapper around the \code{sci_figure} function
to illustrate reproducibility in a two-experiment setting.
}

