% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icons}
\alias{icons}
\title{scifigure icons}
\format{A list of length 44, with each item a 75x75x4 bitmap}
\usage{
icons
}
\description{
A dataset containing icon images used to render all figures
in the scifigure package.
}
\keyword{datasets}
